backend2/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-SocketIO==5.3.4
python-socketio==5.9.0
eventlet==0.33.3
numpy==1.24.3
scipy==1.11.2
matplotlib==3.7.2
plotly==5.17.0
python-dotenv==1.0.0
pyyaml==6.0